feat: ensure product quantity can be decreased from cart

+++ src/actions/drecrease-cart-quantity/index.ts
@@ -2,17 +2,19 @@
 
 import { eq } from "drizzle-orm";
 import { headers } from "next/headers";
-import z from "zod";
 
 import { db } from "@/db";
 import { cartItemTable } from "@/db/schema";
 import { auth } from "@/lib/auth";
 
-import { removeProductFromCartSchema } from "./schema";
-export const removeProductFromCart = async (
-  data: z.infer<typeof removeProductFromCartSchema>,
+import {
+  DecreaseCartProductQuantitySchema,
+  decreaseCartProductQuantitySchema,
+} from "./schema";
+export const decreaseCartProductQuantity = async (
+  data: DecreaseCartProductQuantitySchema,
 ) => {
-  removeProductFromCartSchema.parse(data);
+  decreaseCartProductQuantitySchema.parse(data);
 
   const session = await auth.api.getSession({
     headers: await headers(),
@@ -38,5 +40,15 @@ export const removeProductFromCart = async (
     throw new Error("Unauthorized");
   }
 
-  await db.delete(cartItemTable).where(eq(cartItemTable.id, cartItem.id));
+  if (cartItem.quantity === 1) {
+    await db.delete(cartItemTable).where(eq(cartItemTable.id, cartItem.id));
+    return;
+  }
+
+  await db
+    .update(cartItemTable)
+    .set({
+      quantity: cartItem.quantity - 1,
+    })
+    .where(eq(cartItemTable.id, cartItem.id));
 };
